requirements.txt
fastapi[all]
SQLAlchemy
passlib[bcrypt]
PyJWT
flask-admin